Use uint32 instead of unsigned

+++ foundation/foundation/jzon.h
@@ -5,8 +5,6 @@
 #include <stdbool.h>
 #endif
 
-#include <stdint.h>
-
 #ifdef __cplusplus
 extern "C"
 {
@@ -39,7 +37,7 @@ typedef struct JzonValue
 
 struct JzonKeyValuePair {
 	char* key;
-	uint64_t key_hash;
+	unsigned long long key_hash;
 	JzonValue* value;
 };
